feat:introduce diffent clearance types to execute certain actions

+++ src/admin/manage.admin.ts
@@ -4,6 +4,10 @@ import { loadSidebar, sidebarActions } from '../components/sidebar';
 import { getReportByID } from '../lib/get-reports';
 import { delReport, updateReport } from '../lib/update-report';
 import { loadSpinner, spinnerActionsAdd, spinnerActionsRemove } from '../components/spinner';
+import Endpoints from '../lib/endpoint';
+import { checkDelClearance } from '../lib/check-del-clearance';
+import { popUp, popupActions } from '../components/popup';
+import { getTechnicians } from '../lib/get-technicians';
 
 const managePage = document.querySelector<HTMLDivElement>('#app')!
 const container = document.createElement("div");
@@ -33,6 +37,7 @@ export const loadManagePage = () => {
         <div class="action-btns">
           <button class="btn-update" id="btn-update">Update Status</button>
           <button class="btn-add" id="btn-add">Add Notes</button>
+          <button class="btn-assign" id="btn-assign">Assign A Tech</button>
           <button class="btn-delete" id="btn-delete">Delete Report</button>
         </div>
 
@@ -50,13 +55,13 @@ export const loadManagePage = () => {
 }
 
 const loadDetails = async (detailsID: string) => {
-  const reportIdUrl = `https://nodeserver-v2.onrender.com/api/report/id/${detailsID}`
-  // const reportIdUrl = `http://localhost:8080/api/report/id/${detailsID}`
+  spinnerActionsAdd()
+  const reportIdUrl = Endpoints.reportIdUrl(detailsID)
   const res = await getReportByID(reportIdUrl);
   if (!res?.ok) {
     return
   }
-
+  spinnerActionsRemove();
 
   const report = res?.content.report
   tokenID = report?.tokenID;
@@ -83,6 +88,9 @@ const loadDetails = async (detailsID: string) => {
               <span class="info-key">Room Number</span>|<span class="info-value">${report?.room}</span>
             </div>
             <div class="info-row">
+              <span class="info-key">Technician</span>|<span class="info-value">${report?.technician}</span>
+            </div>
+            <div class="info-row">
               <span class="info-key">Submitted On</span>|<span class="info-value">${report?.submittedOn}</span>
             </div>
   `
@@ -100,6 +108,7 @@ const ticketInfoContainer = document.querySelector(".ticket-info") as HTMLDivEle
 const closeBtn = document.getElementById("btn-close");
 const updateBtn = document.getElementById("btn-update");
 const addBtn = document.getElementById("btn-add");
+const assignBtn = document.getElementById("btn-assign");
 const deleteBtn = document.getElementById("btn-delete");
 
 closeBtn?.addEventListener("click", () => {
@@ -116,9 +125,25 @@ addBtn?.addEventListener("click", (e) => {
   addNotes();
 })
 
+if (adminDetails.clearance_level > 0) {
+  assignBtn?.setAttribute("disabled", "")
+}
+assignBtn?.addEventListener("click", (e) => {
+  e.preventDefault()
+  assignReport()
+})
+
+if (adminDetails.clearance_level > 0) {
+  deleteBtn?.setAttribute("disabled", "")
+}
 deleteBtn?.addEventListener("click", (e) => {
   e.preventDefault()
-  deleteReport()
+  if (checkDelClearance(adminDetails.clearance_level)) {
+    deleteReport()
+  } else {
+    popUp("Forbidden Operation", "This action can only be done by admin")
+    popupActions();
+  }
 })
 
 
@@ -178,8 +203,7 @@ const updateStatus = () => {
   updateBtn?.addEventListener("click", async (e) => {
     e.preventDefault();
     spinnerActionsAdd()
-    const updateReportUrl = `https://nodeserver-v2.onrender.com/api/report/status/${id}`
-    // const updateReportUrl = `http://localhost:8080/api/report/status/${id}`
+    const updateReportUrl = Endpoints.updateReportStatusUrl(id)
     const res = await updateReport(updateReportUrl, { status: statusValue })
     if (!res?.ok) {
       console.log(res?.content)
@@ -229,9 +253,94 @@ const addNotes = () => {
   updateBtn?.addEventListener("click", async (e) => {
     e.preventDefault();
     spinnerActionsAdd()
-    const updateReportUrl = `https://nodeserver-v2.onrender.com/api/report/notes/${id}`
-    // const updateReportUrl = `http://localhost:8080/api/report/notes/${id}`
+    const updateReportUrl = Endpoints.updateReportNotesUrl(id)
     const res = await updateReport(updateReportUrl, { notes: notes.value })
+    if (!res?.ok) {
+      console.log(res?.content)
+    } else {
+      console.log(res?.content)
+    }
+    spinnerActionsRemove()
+    window.location.reload();
+  })
+
+  cancelBtn?.addEventListener("click", (e) => {
+    e.preventDefault()
+    const updateModal = document.querySelector(".update_modal") as HTMLElement;
+    managePage.removeChild(updateModal);
+  })
+}
+
+const assignReport = async () => {
+  const techs = await getTechnicians(Endpoints.techniciansUrl);
+  const updateModal = document.createElement("div");
+  updateModal.classList.add("update_modal")
+  updateModal.innerHTML = `
+    <div>
+      <div class="header">
+        <h1>Assign Report To A Technician</h1>
+        <h2>TokenID: ${tokenID}</h2>
+      </div>
+      <form>
+        <div>
+          <h3>Enter Technician Email</h3>
+          <div class="assign">
+             <input type="text" id="tech-email" name="tech-email" placeholder="Email" required>
+             <ul id="autocomplete"></ul>
+          </div>
+        </div>
+        <div class="action-btns">
+          <button class="btn-status" id="btn-status">Update</button>
+          <button class="btn-close" id="btn-cancel">Cancel</button>
+        </div>
+      </form>
+    </div>
+  `
+  managePage.appendChild(updateModal);
+  const updateBtn = document.getElementById("btn-status");
+  const cancelBtn = document.getElementById("btn-cancel");
+  const email = document.getElementById("tech-email") as HTMLInputElement;
+  const autocomplete = document.getElementById("autocomplete") as HTMLUListElement
+
+  email.addEventListener("keyup", (e) => {
+    e.preventDefault()
+    if (email.value.length < 1) {
+      autocomplete.replaceChildren("")
+      return;
+    }
+    complete(email.value)
+
+  });
+
+  const complete = (val: string) => {
+    let searchList: any = [];
+    techs?.content.forEach((c: any) => {
+
+      if (c.email.toLowerCase().includes(val.toLowerCase())) {
+        searchList.push(c.email)
+      }
+    });
+
+    autocomplete.replaceChildren("")
+    searchList.forEach((l: any) => {
+      autocomplete.innerHTML += `
+      <li class="email" data-set-name=${l}>${l}</li>
+    `
+    })
+    const emails = document.querySelectorAll(".email");
+    emails.forEach((m) => {
+      m.addEventListener("click", () => {
+        email.value = m.getAttribute("data-set-name") || " "
+        autocomplete.replaceChildren("")
+      })
+    })
+  }
+
+  updateBtn?.addEventListener("click", async (e) => {
+    e.preventDefault();
+    spinnerActionsAdd()
+    const updateReportUrl = Endpoints.assignReportUrl(id)
+    const res = await updateReport(updateReportUrl, { technician: email.value })
     if (!res?.ok) {
       console.log(res?.content)
     } else {
@@ -276,9 +385,8 @@ const deleteReport = () => {
   updateBtn?.addEventListener("click", async (e) => {
     e.preventDefault();
     spinnerActionsAdd()
-    const updateReportUrl = `https://nodeserver-v2.onrender.com/api/report/delete/${id}`
-    // const updateReportUrl = `http://localhost:8080/api/report/notes/${id}`
-    const res = await delReport(updateReportUrl)
+    const updateReportUrl = Endpoints.updateReportDeleteUrl(id)
+    const res = await delReport(updateReportUrl, { adminID: adminDetails.email })
 
     if (!res?.ok) {
       console.log(res?.content)
